data_ingestion/requirements.txt
dagster
dagster-cloud
google-cloud-storage
google-cloud-pubsub
pandas
pyyaml
